requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.1
annotated-types==0.7.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.3
black==25.1.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
click==8.1.8
fake-useragent==2.1.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
loguru==0.7.3
lxml==5.3.1
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-mock==3.14.0
python-dotenv==1.0.1
soupsieve==2.6
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzlocal==5.3.1
yarl==1.18.3
